Use pack name in page title for a single pack
When linking to an SWI-Prolog pack from a social networking site, the
site typically displays the link's HTML title as part of the summary.
The old website always showed "SWI-Prolog packages" which makes it look
like the same link is being shared repeatedly.

By including the pack's name in the title, it becomes obvious that we're
linking to a specific pack, not the entire package repository.