Use library(crypto) for computing the hash.
The fact that this is slower for very large files does not really
matter in this case, since the hash is only computed once per file.
It is much more important to stress the functionality of SWI-Prolog,
and this acts as an automatic test case for the OpenSSL bindings
in addition to being shorter and more portable.

This is the first step towards signatures via library(crypto).